+++ packages/carlin/src/generateEnv/generateEnvCommand.ts
@@ -1,5 +1,6 @@
-import { CommandModule, InferredOptionTypes } from 'yargs';
-import { generateEnv } from './generateEnv';
+import type { CommandModule, InferredOptionTypes } from 'yargs';
+
+import { type DeployOutput, generateEnv } from './generateEnv';
 
 export const DEFAULT_ENVIRONMENT = 'Staging';
 
@@ -21,7 +22,9 @@ export const options = {
 export const generateEnvCommand: CommandModule<
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   any,
-  InferredOptionTypes<typeof options>
+  InferredOptionTypes<typeof options> & {
+    envFromDeployOutputs?: DeployOutput[];
+  }
 > = {
   command: ['generate-env', 'ge', 'env'],
   describe: 'Generate environment files.',